+++ resources/locales.clj
@@ -1,5 +1,5 @@
 {
-  :locales  #{"en"}
+  :locales  #{"en" "es" "pt" }
   :packages ["metabase"]
   :bundle   "metabase.Messages"
 }