+++ src/metabase/db/spec.clj
@@ -48,11 +48,11 @@
   (if db
     (merge {:classname "com.stratio.jdbc.core.jdbc4.StratioDriver" ; must be in classpath
             :subprotocol "crossdata"
-            :subname (str "//Server=" host ":" port ";UID=" user ";SSL=true;LogLevel=3;LogPath=/tmp/crossdata-jdbc-logs")}
+            :subname (str "//Server=" host ":" port ";UID=" user ";SSL=true;LogLevel=3;LogPath=/tmp/crossdata-jdbc-logs;TIMEOUT=15")}
            (dissoc opts :host :port :db))
     (merge {:classname "com.stratio.jdbc.core.jdbc4.StratioDriver" ; must be in classpath
             :subprotocol "crossdata"
-            :subname (str "//Server=" host ":" port ";UID=" user ";LogLevel=3;LogPath=/tmp/crossdata-jdbc-logs")}
+            :subname (str "//Server=" host ":" port ";UID=" user ";LogLevel=3;LogPath=/tmp/crossdata-jdbc-logs;TIMEOUT=15")}
            (dissoc opts :host :port :db))))