+++ src/metabase/util/schema.clj
@@ -111,12 +111,24 @@
   (with-api-error-message (s/constrained s/Str (complement str/blank?) "Non-blank string")
     (tru "value must be a non-blank string.")))
 
+(def IntGreaterThanOrEqualToZero
+  "Schema representing an integer than must also be greater than or equal to zero."
+  (with-api-error-message
+   (s/constrained s/Int (partial <= 0) (tru "Integer greater than or equal to zero"))
+   (tru "value must be an integer greater than or equal to zero.")))
+
 ;; TODO - rename this to `PositiveInt`?
 (def IntGreaterThanZero
   "Schema representing an integer than must also be greater than zero."
   (with-api-error-message
       (s/constrained s/Int (partial < 0) (tru "Integer greater than zero"))
     (tru "value must be an integer greater than zero.")))
+
+(def NonNegativeInt
+  "Schema representing an integer 0 or greater"
+  (with-api-error-message
+   (s/constrained s/Int (partial <= 0) (tru "Integer greater than or equal to zero"))
+   (tru "value must be an integer zero or greater.")))
 
 (def PositiveNum
   "Schema representing a numeric value greater than zero. This allows floating point numbers and integers."