+++ testsAT/src/test/java/com/stratio/schema/discovery/installation/DISC_nightly_IT.java
@@ -13,6 +13,7 @@ import org.testng.annotations.Factory;
         "src/test/resources/features/001_installation/000_disc_createPolicies.feature",
         "src/test/resources/features/001_installation/001_disc_installPostgres.feature",
         "src/test/resources/features/001_installation/002_disc_installDiscovery.feature",
+        "src/test/resources/features/001_installation/003_disc_setupDiscovery.feature",
         "src/test/resources/features/002_connections/001_disc_connectionPG.feature",
         "src/test/resources/features/002_connections/002_disc_connectionXD.feature",
         //"src/test/resources/features/004_settings/005_disc_loginUser.feature",